Fixed FormFlow::saveCurrentStepData()
This fixes a bug where "file" fields in an embedded subform replaced all other fields in the same flow step. `array_merge()` simply replaces the data of the embedded form with the file fields of the embedded form. We have to merge recursively to add the file fields in embedded forms.

This is just a quick fix for the problem, we probably need to adjust the test suite to cover this case. The quickest way is probably to modify the test for #64 adding a file field.