Update FormFlow.php
The refine current step is recursively sending the entire array back into its self. Why? it doesnt make sense. 
Whats happening is this. If you have an array of say 3 steps. If say the 1st step is set to be skipped, it will send the entire array back into itself and iterate over all steps again.  This means if you have a form with 10 steps, it may iterate at a worst case 10 + 9 + 8 + 7 ... 1 times..