Update FormFlow.php
Includes a flag to determine if we should expire the flow on a POST/PUT request, so we can create a flow from a POST/PUT request like this workflow

Load Route
We do some stuff, for example in AngularJS
Once finished our stuff, we submit data to the same controller (or another controller).
That controller, handles the data and creates a flow according to submitted data.
Flow is rendered.

This workflow won't work right now, because every POST/PUT request expires the flow.